+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # 🚀Changelog
 
 -------------------------------------------------------------------------------------------------------------
-# 5.8.28(2024-04-24)
+# 5.8.28(2024-04-25)
 
 ### 🐣新特性
 * 【core   】      修正XmlUtil的omitXmlDeclaration描述注释（issue#I9CPC7@Gitee）
@@ -13,6 +13,7 @@
 * 【cache  】      CacheUtil.newTimedCache增加有schedulePruneDelay参数的重载方法（issue#I9HO25@Gitee）
 * 【core   】      NumberChineseFormatter提供阿拉伯转中文支持多位小数的方法（pr#3552@Github）
 * 【captcha】      Captcha.setBackground为null时背景透明（issue#3558@Github）
+* 【captcha】      HttpDownloader.downloadBytes增加超时参数重载（issue#3556@Github）
 
 ### 🐞Bug修复
 * 【http   】      修复HttpUtil.urlWithFormUrlEncoded方法重复编码问题（issue#3536@Github）
@@ -25,6 +26,7 @@
 * 【core   】      修复TemporalAccessorConverter自定义格式转换问题（issue#I9HQQE@Gitee）
 * 【cron   】      修复CronPattern.nextMatchAfter匹配初始值问题（issue#I9FQUA@Gitee）
 * 【core   】      修复FileUtil.copyFile没有创建父目录导致的问题（issue#3557@Github）
+* 【http   】      修复HttpDownloader全局超时无效问题（issue#3556@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.27(2024-03-29)